Move programs to their own app
