+++ corehq/apps/commtrack/management/commands/product_program_last_modified.py
@@ -1,5 +1,5 @@
 from django.core.management.base import BaseCommand
-from corehq.apps.commtrack.models import Program
+from corehq.apps.programs.models import Program
 from corehq.apps.products.models import Product
 from dimagi.utils.couch.database import iter_docs
 from datetime import datetime
