+++ corehq/apps/commtrack/tests/test_fixture.py
@@ -6,8 +6,8 @@ from casexml.apps.case.xml import V1
 from corehq.apps.app_manager.tests.util import TestFileMixin
 from corehq.apps.commtrack.fixtures import program_fixture_generator
 from corehq.apps.products.fixtures import product_fixture_generator
-from corehq.apps.commtrack.models import Program
 from corehq.apps.products.models import Product
+from corehq.apps.programs.models import Program
 from corehq.apps.commtrack.tests.util import CommTrackTest
 from corehq.apps.commtrack.tests.util import bootstrap_user
 from casexml.apps.phone.models import SyncLog
