+++ corehq/apps/products/views.py
@@ -19,7 +19,7 @@ from corehq.apps.products.models import Product
 from corehq.apps.products.forms import ProductForm
 from corehq.apps.commtrack.views import BaseCommTrackManageView
 from corehq.apps.commtrack.util import encode_if_needed
-from corehq.apps.commtrack.models import Program
+from corehq.apps.programs.models import Program
 from corehq.apps.custom_data_fields.models import CustomDataFieldsDefinition
 from corehq.apps.custom_data_fields.views import (
     CustomDataEditor, CustomDataFieldsMixin
