+++ corehq/apps/settings/urls.py
@@ -4,6 +4,7 @@ from corehq.apps.domain.urls import domain_settings
 from corehq.apps.cloudcare.urls import settings_urls as cloudcare_settings
 from corehq.apps.commtrack.urls import settings_urls as commtrack_settings
 from corehq.apps.products.urls import settings_urls as product_settings
+from corehq.apps.programs.urls import settings_urls as program_settings
 from corehq.apps.facilities.urls import settings_urls as facilities_settings
 from corehq.apps.locations.urls import settings_urls as location_settings
 
@@ -25,6 +26,7 @@ domain_specific = patterns('',
     (r'^cloudcare/', include(cloudcare_settings)),
     (r'^commtrack/', include(commtrack_settings)),
     (r'^products/', include(product_settings)),
+    (r'^programs/', include(program_settings)),
     (r'^locations/', include(location_settings)),
     (r'^facilities/', include(facilities_settings)),
     url(r'^api/id_mapping/$', 'corehq.apps.settings.views.project_id_mapping', name="project_id_mapping")
