+++ corehq/apps/users/forms.py
@@ -18,7 +18,8 @@ from corehq.apps.registration.utils import handle_changed_mailchimp_email
 from corehq.apps.users.models import CouchUser
 from corehq.apps.users.util import format_username
 from corehq.apps.app_manager.models import validate_lang
-from corehq.apps.commtrack.models import CommTrackUser, Program, SupplyPointCase
+from corehq.apps.commtrack.models import CommTrackUser, SupplyPointCase
+from corehq.apps.programs.models import Program
 from bootstrap3_crispy import layout as cb3_layout
 from bootstrap3_crispy import helper as cb3_helper
 import re
