+++ settings.py
@@ -226,6 +226,7 @@ HQ_APPS = (
     'corehq.apps.loadtestendpoints',
     'corehq.apps.locations',
     'corehq.apps.products',
+    'corehq.apps.programs',
     'corehq.apps.commtrack',
     'corehq.apps.consumption',
     'couchforms',
@@ -960,6 +961,7 @@ COUCHDB_APPS = [
     'pillowtop',
     'pillow_retry',
     'products',
+    'programs',
     'reminders',
     'reports',
     'sofabed',
